test(ApplicationTest): 성공 테스트와 실패 테스트 클래스 분리
