+++ src/test/java/racingcar/ApplicationFailTest.java
@@ -4,40 +4,12 @@ import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static racingcar.message.ErrorMessage.BLANK_INPUT_ERROR;
 import static racingcar.message.ErrorMessage.CONTAINS_SPACE_ERROR;
 
-class ApplicationTest extends NsTest {
-    private static final int MOVING_FORWARD = 4;
-    private static final int STOP = 3;
-
-    @DisplayName("통합 기능 테스트")
-    @Test
-    void 기능_테스트() {
-        assertRandomNumberInRangeTest(
-                () -> {
-                    run("pobi,woni", "1");
-                    assertThat(output()).contains("pobi : -", "woni : ", "최종 우승자 : pobi");
-                },
-                MOVING_FORWARD, STOP
-        );
-    }
-
-    @DisplayName("이동 시도 횟수가 0번이어도 잘 작동해야 한다")
-    @Test
-    void 이동횟수는_0_이상이어야_한다() {
-        assertRandomNumberInRangeTest(
-                () -> {
-                    run("pobi,woni", "0");
-                    assertThat(output()).contains("최종 우승자 : pobi,woni");
-                },
-                MOVING_FORWARD, STOP
-        );
-    }
+public class ApplicationFailTest extends NsTest {
 
     @DisplayName("이름은 6글자 이상이면 예외가 발생한다.")
     @Test
